Rename WindowGrabber project to WindowsInteraction
