+++ nardnob.InputTracker-Winforms.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.10.35004.147
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "nardnob.InputTracker.WinForms", "WinForms\nardnob.InputTracker.WinForms.csproj", "{C64C3659-D381-41C0-AB05-3C327779A27A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "nardnob.InputTracker.WindowGrabber", "WindowGrabber\nardnob.InputTracker.WindowGrabber.csproj", "{B49D5F5C-F43F-492C-A9E1-63CD78BDA1E2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "nardnob.InputTracker.WindowsInteraction", "WindowsInteraction\nardnob.InputTracker.WindowsInteraction.csproj", "{B49D5F5C-F43F-492C-A9E1-63CD78BDA1E2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
